V4_SOVEREIGN_MAGIC/CODE/requirements.txt
# V4_SOVEREIGN_MAGIC - Python Dependencies
# Phase 1 RAG Upgrades (2026-03-05)
# Updated: Using sentence-transformers for local GPU embeddings

# Core Dependencies
marker-pdf==1.10.2
supabase
openai
python-dotenv
fastapi
requests

# Phase 1 RAG Upgrades
tiktoken>=0.5.0           # Token counting for chunking
rank-bm25>=0.2.2          # BM25 for hybrid search
sentence-transformers>=2.2.0  # Local GPU embeddings (no API costs)

# Optional: For local development/testing
# pgvector>=0.2.0         # PostgreSQL vector extension (installed via Supabase)
# psycopg2-binary         # For direct database connections
